Fix integer division in calculating average times

+++ PartA/fcfs.c
@@ -56,7 +56,7 @@ void first_come_first_served(struct task_t* task, int size)
 
 float calculate_average_wait_time(struct task_t* task, int size)
 {
-    int totalTime = 0;
+    float totalTime = 0;
 
     for(int i = 0; i < size; i++)
     {
@@ -69,7 +69,7 @@ float calculate_average_wait_time(struct task_t* task, int size)
 
 float calculate_average_turn_around_time(struct task_t* task, int size)
 {
-    int totalTime = 0;
+    float totalTime = 0;
 
     for(int i = 0; i < size; i++)
     {
